requirements.txt
python-telegram-bot==21.5
motor==3.5.1
pymongo==4.8.0
aiohttp==3.10.5
python-dotenv==1.0.1
